% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PSigA.R
\docType{data}
\name{AML_meta}
\alias{AML_meta}
\title{AML dataset metadata}
\format{A data frame with 424 rows (patients) and 3 columns:
\itemize{
    \item karyotype: Cytogenetic aberration
    \item gender: 1 = male, 0 = female
    \item age: patient's age in years
}}
\source{
\url{https://tcga-data.nci.nih.gov/tcga/}

\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE15434}

\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE13159}

\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE14468}
}
\usage{
data(AML)
}
\description{
Clinical data for the \code{\link{AML}} dataset.
}
\keyword{datasets}

