% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PSigA.R
\docType{data}
\name{blood}
\alias{blood}
\title{Bone marrow samples from 60 healthy individuals}
\format{A data frame with 54675 rows (probes) and 60 columns. The column
names describe the blood cell type of the respective sample.}
\usage{
data(blood)
}
\description{
Gene expression data of 60 FACS sorted blood samples using the Affymetrix
HGU 133 plus2 platform.
}
\keyword{datasets}

