% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PSigA.R
\docType{data}
\name{DSigDB}
\alias{DSigDB}
\title{Drug SIGnature DataBase}
\format{An object of type list with 22,528 entries, each one representing a
gene signature. The name of each list entry represents the signature's name.
Additionally, \code{MSigDB} entries contain the following collection
identifiers:
\itemize{
 \item{D1: FDA Approved (1202)}
 \item{D2: Kinase Inhibitors (1220)}
 \item{D3: Perturbagen Signatures (1998)}
 \item{D4: Computational Drug Signatures (18107)}
}}
\source{
\url{http://tanlab.ucdenver.edu/DSigDB/DSigDBv1.0/}
}
\usage{
data(DSigDb)
}
\description{
22,528 gene pathways obtained from Drug SIGnature DataBase (DSigDB v1.0).
}
\keyword{datasets}

