% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{gene2probe}
\alias{gene2probe}
\title{Convert gene identifiers to microarray probeID}
\usage{
gene2probe(genes, db, keytype = "SYMBOL")
}
\arguments{
\item{genes}{List of gene identifiers to be converted}

\item{db}{Microarray platform. See details for a list of valid inputs}

\item{keytype}{the keytype that matches the keys used in the \code{gene} parameter. See a list with valid inputs
in the details below. (Default = "SYMBOL")}
}
\value{
keys a data frame of possible values
}
\description{
\code{gene2probe} returns probeIDs compatible with several Microarray Platforms
}

