% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/measureLoadings.R
\name{measureLoadings}
\alias{measureLoadings}
\title{Measure variable loadings}
\usage{
measureLoadings(pca, varCutoff = 0.75, pcs = NULL)
}
\arguments{
\item{pca}{\code{\link{prcomp}} object.}

\item{varCutoff}{cumulative variance cutoff.}

\item{pcs}{when set the function will compute the Euclidean norm for
selected principal components instead of using the first \code{n}
(overrides \code{varCutoff}). Vector of length 2.}
}
\value{
norms a vector with the Euclidean Norm of all the loadings.
}
\description{
Measures the Euclidean Norm of the PCA loadings (rotation
matrix) of the first \code{n} principal components, where the cumulative
variance on the \code{n}-th principal component is above a cutoff.
}
\author{
Nikos Sidiropoulos
}

