% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/measureSpread.R
\name{measureSpread}
\alias{measureSpread}
\title{Measures the area under the density curve on PC space}
\usage{
measureSpread(sigpca, varCutoff = 0.75, denCutoffLow = 0.005,
  denCutoffHigh = 0.05, pc = NULL, sigName = "NA")
}
\arguments{
\item{sigpca}{signature PCA. Object of class \code{\link{prcomp}}. Can be
produced with \code{\link{signaturePCA}}.}

\item{varCutoff}{cumulative variance cutoff. The function will measure the
area under the components that describe more or equal portion of the
variance defined by this parameter.}

\item{denCutoffLow}{lower density cutoff.}

\item{denCutoffHigh}{upper density cutoff.}

\item{pc}{when set the function will compute the area in the first \code{pcs}
components instead of using the cumulative variance percentage.
(overrides \code{varCutoff}).}

\item{sigName}{?}
}
\value{
a data frame with columns: Signature Name, Number of principal
components used, Density area.
}
\description{
\code{measureSpread} integrates the density curves of the first
\emph{n} principle components of a \code{\link{signaturePCA}}, where \emph{n}
is component where the cumulative variance is above a user specified cutoff.
Much more to be written...
}
\author{
Nikos Sidiropoulos
}

