% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/measureSpreadParallel.R
\name{measureSpreadParallel}
\alias{measureSpreadParallel}
\title{Parallel execution of \code{\link{measureSpread}}}
\usage{
measureSpreadParallel(data, signatures, db, threads = 4, ...)
}
\arguments{
\item{data}{gene expression matrix}

\item{signatures}{list of gene signatures}

\item{threads}{number of cpu threads to be used}

\item{...}{arugments passed to \code{\link{measureSpread}}}
}
\value{
a data frame with columns: Signature Name, Number of principal
components used, Density area.
}
\description{
\code{measureSpreadParallel}
}
\author{
Nikos Sidiropoulos
}

