% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SigPCA.R
\docType{data}
\name{MSigDB}
\alias{MSigDB}
\title{Molecular Signature DataBase}
\format{An object of type list with 10,348 entries, each one representing a
gene signature. The name of each list entry represents the signature's name.
Additionally, \code{MSigDB} entries contain the following collection
identifiers:
\itemize{
 \item{H: hallmark gene sets (50)}
 \item{C1: positional gene sets (326)}
 \item{C2: curated gene sets (4725)}
 \item{C3: motif gene sets (836)}
 \item{C4: computational gene sets (858)}
 \item{C5: GO gene sets (1454)}
 \item{C6: oncogenic signatures (189)}
 \item{C7: immunologic signatures (1910)}
}}
\source{
\url{http://www.broadinstitute.org/gsea/msigdb/index.jsp}
}
\usage{
data(MSigDb)
}
\description{
10,348 gene pathways obtained from Molecular Signature Database
(MSigDB v5.0).
}
\keyword{datasets}

