% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseData.R
\name{parseData}
\alias{parseData}
\title{Parse data to be used \code{\link{PsigA}} functions}
\usage{
parseData(data, geneIds)
}
\arguments{
\item{data}{a data frame of matrix with gene expression values where rows
represent genes and columns represent samples.}

\item{geneIds}{vector of length \code{nrow(data)} containing gene names in
of the desired format. The format must be the same as the format of the
signatures that will be used in the \code{PsigA} functions.}
}
\value{
An object of class \code{data} where the rows correspond to
\code{unique(geneIds)}.
}
\description{
The function converts the gene identifiers (rownames) of a gene
expression matrix to a desired format, supplied by the parameter
\code{geneIds}. If any of the gene identifiers correspond to more than one
row of the expression matrix (e.g. probes that map to the same gene), the
median expression of each patient..
}
\examples{

require(Biobase)
require(breastCancerVDX)

data(vdx)

#use fData() to get the gene symbols for vdx.
VDXparsed <- parseData(data = exprs(vdx), geneIds = fData(vdx)$Gene.symbol)


require(ALL)
data(ALL)

ALLexprs <- exprs(ALL)

#featureData ("fData()") for ALL are missing so we need to get the gene
#symbols manually.

annotation(ALL)

require(hgu95av2.db)

keys <- AnnotationDbi::select(hgu95av2.db, rownames(ALLexprs),
                             "SYMBOL", "PROBEID")
#remove probe duplicates
geneIds <- keys$SYMBOL[ !duplicated(keys$PROBEID) ]

ALLparsed <- parseData(ALLexprs, geneIds)

}

