% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakDistance2d.R
\name{peakDistance2d}
\alias{peakDistance2d}
\title{Peak distance in a 2D signature-PCA}
\usage{
peakDistance2d(signature, data, threshold = 0.005, n = 200,
  magnitude = FALSE, scale = FALSE, filtered = FALSE)
}
\arguments{
\item{signature}{character vector with the signature's gene identifiers}

\item{data}{Gene expression matrix where rownames correspond to unique gene
identifiers in \code{signature} fortmat and columns correspond to samples.}

\item{threshold}{density cutoff. Density values lower than the
\code{threshold} will not be considered peaks. Useful when outliers are
present in the PCA.}

\item{n}{Number of grid points in each direction. Can be scalar or a
length-2 integer vector. See \link{kde2d}.}

\item{magnitude}{When TRUE the score is multiplied by the cluster density.
Default: FALSE.}

\item{scale}{a logical value indicating whether the variables should be
scaled to have unit variance before the analysis takes place.
See prcomp for more details. Default: FALSE.}

\item{filtered}{logical value indicating if genes in the supplied
\code{signature} list that are not present in the \code{data} have been
filtered out. Default: FALSE.}
}
\value{
A vector of length 2. The first value corresponds to the score and
the second to the number of genes in the \code{signature} that were found in
the \code{data}.
}
\description{
Measures the Euclidean distance between the two
highest peaks of the 2D density function. The density is estimated from the
samples' distribution in the first two principal components of a PCA, using
the genes of a given signature.
}
\examples{

dummyData <- do.call(rbind, lapply(seq(0.1, 0.9, by = 0.1),
                     rnorm, n = 100, m = 6))

#add a row with bimodal gene expression
dummyData <- rbind(dummyData, c(rnorm(70, 6, 0.1), rnorm(30, 9, 0.1)))

rownames(dummyData) <- paste(rep("gene", nrow(dummyData)),
                             seq(1, nrow(dummyData)), sep = "")
rownames(dummyData)

dummySig <- c("gene1", "gene8", "gene9", "gene10", "gene20", "gene30")

peakDistance2d(dummySig, dummyData)

#values correspond to the peak distance, the number of genes from the
#signature found in the data and the total number of genes in the signature
#respectively

#removing the bimodal gene from the signature results to a lower score
peakDistance2d(dummySig[-4], dummyData)

}

