% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{probe2geneMap}
\alias{probe2geneMap}
\title{Converts probe id to gene symbols}
\usage{
probe2geneMap(keys, db)
}
\arguments{
\item{keys}{a character vector of probe identifiers}

\item{db}{microarray platform}
}
\value{
character vector of gene symbols
}
\description{
Converts probe id to gene symbols
}
\author{
Nikos Sidiropoulos
}

