% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSigA.R
\docType{package}
\name{PSigA}
\alias{PSigA}
\alias{PSigA-package}
\title{PSigA: A gene-signature ranking method based on sample density in PCA space.}
\description{
The PSigA package provides a set of functions to perform Principal Component
analysis based on genetic pathways. Moreover PSigA incorporates an algorithm
to measure the spread of the data in the PCA space and rank the
gene-signatures based on their ability to split the data in distinct
entities.
}

