% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SigPCA.R
\docType{data}
\name{RAPIN}
\alias{RAPIN}
\title{Rapin et al. signatures}
\format{An object of type list with 24 entries, each one representing a
gene signature. The name of each list entry represents the signature's name.}
\source{
\url{http://www.bloodjournal.org/content/123/6/894}
}
\usage{
data(RAPIN)
}
\description{
24 gene pathways obtained from Rapin et al. (2014).
}
\keyword{datasets}

