% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scoreSigs.R
\name{scoreSigs}
\alias{scoreSigs}
\title{Score gene signatures based on PCA density}
\usage{
scoreSigs(data, parsed = FALSE, genes = (if (parsed) NULL), signatures,
  threshold = 0.005, n = 200)
}
\arguments{
\item{data}{data frame of matrix with gene expression values where rows
represent genes and columns represent samples.}

\item{parsed}{logical.}

\item{genes}{vector of length \code{nrow(data)} containing gene names in
"HGNC" format. If \code{parsed = TRUE} the parameter is omitted.}

\item{signatures}{list where every entry is a character vector of
the gene names in HGNC format that correspond to a given gene-pathway.}

\item{threshold}{low density cutoff. Used in \code{\link{peakDistance2d}}.}

\item{n}{Number of grid points in each direction. Can be scalar or a
length-2 integer vector. See \link{kde2d}.}
}
\value{
A data frame with 2 columns; 1) \code{score}, the \code{PSigA}
score of a given signature and 2) \code{size}, the number of genes in the
\code{signature} that were found in the \code{data}. If
\code{parsed = FALSE}, the function returns a list, where the first entry
contains the data frame described above and the second entry the parsed
\code{data}.
}
\description{
Score gene signatures based on PCA density
}
\examples{
require(Biobase)
require(breastCancerVDX)
data(vdx)
data(RAPIN)

VDX <- readSamples(data = exprs(vdx), genes = fData(vdx)$Gene.symbol)

scores <- scoreSigs(VDX, parsed = TRUE, signatures = MSigDB[1827:1838],
                    threshold = 0.003)
head(scores)

#plot top signature
sigPCA <- signaturePCA(MSigDB[[rownames(scores)[1]]], VDX)
sigPlot(sigPCA)
sigBiplot(sigPCA, factor(pData(vdx)$grade), main = rownames(scores)[1])
}

