% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{sigBiplot}
\alias{sigBiplot}
\title{Signature Biplot for principal compontents}
\usage{
sigBiplot(sigPCA, groups = NULL, labels = NULL, pcs = c(1, 2),
  main = "", obs.size = 2, var.size = 3, var.scaled = FALSE,
  palette = "Paired", ...)
}
\arguments{
\item{sigPCA}{a \code{\link{prcomp}} object, preferrably produced using
\code{\link{signaturePCA}}}

\item{groups}{optional factor variable indicating the groups that the
observations belong to. If provided the points will be colored accordingly}

\item{labels}{optional vector of labels for the observations}

\item{pcs}{which PCs to plot. default: 1,2}

\item{main}{plot title}

\item{obs.size}{size of the points for the observations}

\item{var.size}{size of the text for the variables}

\item{var.scaled}{logical value. When set to TRUE the variable text size is
proportional to the Euclidean Norm of the variable's loading
(see \code{\link{measureLoadings}}).}

\item{palette}{colorbrewer palette scheme to be used. Applicable only when
\code{groups} are provided.}

\item{...}{methods passed to \code{\link{ggbiplot}}.}
}
\value{
A ggplot2 plot
}
\description{
Produces a bibplot of the selected signature's PCA. Based on
\code{\link{ggbiplot}} and \code{\link{ggplot2}}.
}
\examples{
data(AML)
data(MSigDB)

sigPCA <- signaturePCA(MSigDB[["HOMOPHILIC_CELL_ADHESION"]], AML)

sigBiplot(sigPCA, groups = AML_meta$karyotype)
sigBiplot(sigPCA, groups = AML_meta$karyotype, labels = AML_meta$karyotype,
          obs.size = 3)
}

