% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SigPCA.R
\docType{data}
\name{sigdb}
\alias{sigdb}
\title{12,617 gene signatures}
\format{A list with 12617 entries, each one representing a gene signature. The name of each list entry represents
the signature's name.}
\source{
\url{http://www.broadinstitute.org/gsea/msigdb/index.jsp}

\url{https://www.broadinstitute.org/cmap/}

\url{http://www.bloodjournal.org/cgi/pmidlookup?view=long&pmid=24363398}
}
\usage{
data(sigdb)
}
\description{
Gene pathways were obtained from Molecular Signature Database (MSigDB), Connectivity Map (cmap)
and the signatures by Rapin et. al.
}
\keyword{datasets}

