% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SigPCA.R
\docType{package}
\name{SigPCA}
\alias{SigPCA}
\alias{SigPCA-package}
\title{SigPCA: A gene-signature based exploratory tool for gene expression data.}
\description{
The SigPCA package provides a set of functions to perform Principal Component analysis based on genetic pathways.
Moreover SigPCA incorporates an algorithm to measure the spread of the data in the PCA space and rank the
gene-signatures based on their ability to split the data in distinct entities.
}

