% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{sigPlot}
\alias{sigPlot}
\title{Signature PCA plot}
\usage{
sigPlot(data, signature, groups = NULL, text = FALSE, pcs = c(1, 2),
  main = "", palette = "Paired", ...)
}
\arguments{
\item{data}{gene expression matrix where rownames correspond to unique gene
identifiers in \code{signature} format, and columns correspond to samples.}

\item{signature}{character vector containing the signature's gene
identifiers.}

\item{groups}{optional factor variable indicating the groups that the
observations belong to.}

\item{text}{when TRUE it plots textual annotations instead of points
based on \code{groups} parameter.}

\item{pcs}{which PCs to plot. default: 1,2}

\item{main}{plot title}

\item{palette}{colorbrewer palette scheme to be used.}

\item{...}{methods passed to \code{\link{ggplot}}}
}
\value{
A ggplot2 plot
}
\description{
Plot of the selected signature's PCA using
\code{\link{ggplot2}}.
}
\examples{
require(Biobase)
require(breastCancerVDX)
data(vdx)

VDX <- parseData(data = exprs(vdx)[5000:9000,],
                 geneIds = fData(vdx)$Gene.symbol[5000:9000])

dummySig <- c("AGR2", "SCGB1D2", "SCGB2A2")

sigPlot(VDX, dummySig)
sigPlot(VDX, dummySig, groups = factor(pData(vdx)$grade), main = "dummySig")
}

