% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PSigA.R
\docType{data}
\name{AML}
\alias{AML}
\title{Microarray data of 183 AML patients}
\format{A data frame with 54675 rows (Affy HGU 133 Plus 2 probes) and 183
columns (patients).}
\source{
\url{https://tcga-data.nci.nih.gov/tcga/}
}
\usage{
data(AML)
}
\description{
Gene expression data of 183 AML patients, retrieved from TCGA, normalized
and batch corrected. Rows represent microarray probes of the Affymetrix HGU
133 Plus2 platform. Each column represents an AML patient.
}
\keyword{datasets}

