% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SigPCA.R
\docType{data}
\name{AML_meta}
\alias{AML_meta}
\title{AML dataset metadata}
\format{A data frame with 183 rows (patients) and 5 variables
\itemize{
    \item karyotype: Cytogenetic aberration
    \item gender: 1 = male, 0 = female
    \item age: patient's age in years
    \item OS: Survival from diagnosis until the last follow up in days
    \item EventOS: Patient status by the time of the last follow up.
          1 = dead, 0 = alive
}}
\source{
\url{https://tcga-data.nci.nih.gov/tcga/}
}
\usage{
data(AML_meta)
}
\description{
Clinical data of the TCGA AML patients.
}
\keyword{datasets}

