% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/peakDistance2d.R
\name{peakDistance2d}
\alias{peakDistance2d}
\title{Peak distance in a 2D signature PCA}
\usage{
peakDistance2d(signature, data, threshold, n = 200)
}
\arguments{
\item{signature}{character vector with the signature's gene identifiers in
HGNC format.}

\item{data}{Gene expression matrix where rownames correspond to unique gene
identifiers (HGNC format \link{http://www.genenames.org}) and columns
correspond to samples.}

\item{threshold}{density cutoff. Density values lower than the
\code{threshold} will not be considered peaks. Usefull when outliers are
present in the PCA.}

\item{n}{Number of grid points in each direction. Can be scalar or a
length-2 integer vector. See \link{kde2d}.}
}
\value{
A vector of length 2. The first value corresponds to the score and
the second to the number of genes in the \code{signature} that were found in
the \code{data}.
}
\description{
Locates and measures the Euclidean distance between the two
highest peaks of the 2D density function, estimated from the samples'
distribution in the first two principal components in a signature PCA
(see \code{\link{signaturePCA}}).
}

