% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakDistance2d.R
\name{peakDistance2d}
\alias{peakDistance2d}
\title{Peak distance in a 2D signature-PCA}
\usage{
peakDistance2d(signature, data, threshold = 0.005, n = 200,
  magnitude = FALSE)
}
\arguments{
\item{signature}{character vector with the signature's gene identifiers}

\item{data}{Gene expression matrix where rownames correspond to unique gene
identifiers in \code{signature} fortmat and columns correspond to samples.}

\item{threshold}{density cutoff. Density values lower than the
\code{threshold} will not be considered peaks. Useful when outliers are
present in the PCA.}

\item{n}{Number of grid points in each direction. Can be scalar or a
length-2 integer vector. See \link{kde2d}.}

\item{magnitude}{When TRUE the score is multiplied by the cluster density.
Default: FALSE.}
}
\value{
A vector of length 2. The first value corresponds to the score and
the second to the number of genes in the \code{signature} that were found in
the \code{data}.
}
\description{
Measures the Euclidean distance between the two
highest peaks of the 2D density function. The density is estimated from the
samples' distribution in the first two principal components of a PCA, using
the genes of a given signature.
}
\examples{

require(Biobase)
require(breastCancerVDX)
data(vdx)
data(RAPIN)

#get the first 5000 probes of the vdx array
VDX <- parseData(data = exprs(vdx)[1:5000,],
                   geneIds = fData(vdx)$Gene.symbol[1:5000])

dummySig <- rownames(VDX)[1:50]

peakDistance2d(dummySig, VDX)

}

