% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{rankedGSEA}
\alias{rankedGSEA}
\title{PCA-based Gene Set Enrichment Analysis (GSEA)}
\usage{
rankedGSEA(data, signatures, p.adj = p.adjust.methods, pcs = c(1, 2))
}
\arguments{
\item{data}{Gene expression matrix where rownames correspond to unique gene
identifiers in \code{signature} format and columns correspond to samples.}

\item{signatures}{list where every entry is a character vector of
the gene names that correspond to a gene-pathway. If the gene format is
different from the one in \code{rownames(data)}, use \code{\link{parseData}}
first.}

\item{p.adj}{p-value correction method. See \code{\link{p.adjust}}.}
}
\description{
...
}

