% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readSamples.R
\name{readSamples}
\alias{readSamples}
\title{Read and samples to be used in \link{PSigA} functions}
\usage{
readSamples(data, genes)
}
\arguments{
\item{data}{data frame of matrix with gene expression values where rows
represent genes and columns represent samples.}

\item{genes}{vector of length \code{nrow(data)} containing gene names in
"HGNC" format.}
}
\value{
A data frame ...
}
\description{
The function ...
}

