% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scoreSigs.R
\name{scoreSigs}
\alias{scoreSigs}
\title{Score gene signatures based on PCA density}
\usage{
scoreSigs(data, parsed = FALSE, genes = (if (parsed) NULL), signatures,
  varCutoff = 0.75, denCutoffLow = 0.005, denCutoffHigh = 0.02,
  pc = NULL, show.all = FALSE)
}
\arguments{
\item{data}{data frame of matrix with gene expression values where rows
represent genes and columns represent samples.}

\item{parsed}{logical.}

\item{pcs}{when set the function will compute the Euclidean norm for
selected principal components instead of using the first \code{n}
(overrides \code{varCutoff}). Vector of length 2.}
}
\value{
norms a vector with the Euclidean Norm of all the loadings.
}
\description{
Score gene signatures based on PCA density
}

