% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreSigs.R
\name{scoreSigs}
\alias{scoreSigs}
\title{Score gene signatures based on PCA density}
\usage{
scoreSigs(data, signatures, threshold = 0.005, n = 200, magnitude = FALSE)
}
\arguments{
\item{data}{data frame of matrix with gene expression values where rows
represent genes and columns represent samples.}

\item{signatures}{list where every entry is a character vector of
the gene names that correspond to a gene-pathway. If the gene format is
different from the one in \code{rownames(data)}, use \code{\link{parseData}}
first.}

\item{threshold}{low density cutoff. Used in \code{\link{peakDistance2d}}.}

\item{n}{Number of grid points in each direction. Can be scalar or a
length-2 integer vector. See \link{kde2d}.}

\item{magnitude}{When TRUE the score is multiplied by the cluster density.
Default: FALSE.}
}
\value{
A data frame with 2 columns; 1) \code{score}, the \code{PSigA}
score of a given signature and 2) \code{size}, the number of genes in the
\code{signature} that were found in the \code{data}. If
\code{parsed = FALSE}, the function returns a list, where the first entry
contains the data frame described above and the second entry the parsed
\code{data}.
}
\examples{

require(Biobase)
require(breastCancerVDX)
data(vdx)
data(RAPIN)

VDX <- parseData(data = exprs(vdx)[1:5000,],
                 geneIds = fData(vdx)$Gene.symbol[1:5000])

scores <- scoreSigs(VDX, RAPIN, threshold = 0.003)
head(scores)

#plot top signature
sigPlot(VDX, RAPIN[[rownames(scores)[1]]])
sigBiplot(VDX, RAPIN[[rownames(scores)[1]]], factor(pData(vdx)$grade),
          main = rownames(scores)[1])

}

