% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{sigBiplot}
\alias{sigBiplot}
\title{Signature Biplot for principal compontents}
\usage{
sigBiplot(data, signature, groups = NULL, labels = NULL, pcs = c(1, 2),
  main = "", obs.size = 2, var.size = 3, var.scaled = FALSE,
  palette = "Paired", ...)
}
\arguments{
\item{data}{gene expression matrix where rownames correspond to unique gene
identifiers in \code{signature} format, and columns correspond to samples.}

\item{signature}{character vector containing the signature's gene
identifiers.}

\item{groups}{optional factor variable indicating the groups that the
observations belong to. If provided the points will be colored accordingly}

\item{labels}{optional vector of labels for the observations}

\item{pcs}{which PCs to plot. default: 1,2}

\item{main}{plot title}

\item{obs.size}{size of the points for the observations}

\item{var.size}{size of the text for the variables}

\item{var.scaled}{logical value. When set to TRUE the variable text size is
proportional to the Euclidean Norm of the variable's loading
(see \code{\link{measureLoadings}}).}

\item{palette}{colorbrewer palette scheme to be used. Applicable only when
\code{groups} are provided.}

\item{...}{methods passed to \code{\link{ggbiplot}}.}
}
\value{
A ggplot2 plot
}
\description{
Produces a bibplot of the selected signature's PCA. Based on
\code{\link{ggbiplot}} and \code{ggplot2}.
}
\examples{

dummyData <- do.call(rbind, lapply(seq(0.1, 0.9, by = 0.1),
                     rnorm, n = 100, m = 6))
#add a row with bimodal gene expression
dummyData <- rbind(dummyData, c(rnorm(70, 6, 0.1), rnorm(30, 9, 0.1)))

rownames(dummyData) <- paste(rep("gene", nrow(dummyData)),
                             seq(1, nrow(dummyData)), sep = "")

dummySig <- c("gene1", "gene8", "gene9", "gene10")

sigBiplot(dummyData, dummySig)

groups <- factor(c(rep("GroupA", 70), rep("GroupB", 30)))
sigBiplot(dummyData, dummySig, groups, main = "dummySig")

}

