% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{sigBiplot}
\alias{sigBiplot}
\title{Produces a bibplot of the selected signature's PCA}
\usage{
sigBiplot(sigPCA, pcs = c(1, 2), groups = NULL, genenames = TRUE,
  main = "Signature", db = "hgu133plus2.db", palette = "Paired", ...)
}
\arguments{
\item{sigPCA}{a \code{\link{prcomp}} object, preferrably produced using \code{\link{signaturePCA}}}

\item{pcs}{which PCs to plot. default: 1,2}

\item{groups}{optional factor variable indicating the groups that the observations belong to.
If provided the points will be colored according to groups}

\item{genenames}{when TRUE, gene symbols are used instead of probe ids as variable names}

\item{main}{plot title}

\item{db}{microarray platform}

\item{palette}{colorbrewer palette scheme to be used. Applicable only when groups are provided.}

\item{...}{methods passed to \code{\link{ggbiplot}}}
}
\value{
signature PCA biplot
}
\description{
To do
}
\author{
Nikos Sidiropoulos
}

