% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{sigBiplot}
\alias{sigBiplot}
\title{Produces a bibplot of the selected signature's PCA}
\usage{
sigBiplot(sigPCA, pcs = c(1, 2), groups = NULL, genenames = TRUE,
  labels = NULL, db = "hgu133plus2.db", main = "Signature",
  obs.size = 2, var.size = 3, var.scaled = FALSE, palette = "Paired",
  ...)
}
\arguments{
\item{sigPCA}{a \code{\link{prcomp}} object, preferrably produced using
\code{\link{signaturePCA}}}

\item{pcs}{which PCs to plot. default: 1,2}

\item{groups}{optional factor variable indicating the groups that the
observations belong to. If provided the points will be colored accordingly}

\item{genenames}{when TRUE, gene symbols are used instead of probe ids as
variable names}

\item{labels}{optional vector of labels for the observations}

\item{db}{microarray platform}

\item{main}{plot title}

\item{obs.size}{size of the points for the observations}

\item{var.size}{size of the text for the variables}

\item{var.scaled}{logical value. When set to TRUE the variable text size is
proportional to the Euclidean Norm of the variable's loading
(see \code{\link{measureLoadings}}).}

\item{palette}{colorbrewer palette scheme to be used. Applicable only when
\code{groups} are provided.}

\item{...}{methods passed to \code{\link{ggbiplot}}}
}
\value{
signature PCA biplot
}
\description{
To do
}
\author{
Nikos Sidiropoulos
}

