% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{signaturePCA}
\alias{signaturePCA}
\title{PCA based on a given gene-signature}
\usage{
signaturePCA(genes, data, db, ...)
}
\arguments{
\item{genes}{Character vector with the gene identifiers}

\item{data}{Data matrix with gene expression values}

\item{db}{Microarray Platform}

\item{...}{arguments passed to \code{\link{prcomp}}.}
}
\description{
\code{signaturePCA} performs principal component analysis on the given data matrix and gene-signature and
returns the results as an object of class \code{prcomp}.
}

