% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{signaturePCA}
\alias{signaturePCA}
\title{PCA based on a given gene-signature}
\usage{
signaturePCA(signature, data, center = TRUE, scale = FALSE, ...)
}
\arguments{
\item{signature}{character vector with the signature's gene identifiers.}

\item{data}{Gene expression matrix where rownames correspond to unique gene
identifiers in \code{signature} format and columns correspond to samples.}

\item{center}{a logical value indicating whether the variables should be
shifted to be zero centered. See \code{\link{prcomp}} for more details.}

\item{scale}{a logical value indicating whether the variables should be
scaled to have unit variance before the analysis takes place. See
\code{prcomp} for more details.}

\item{...}{arguments passed to \code{prcomp}.}
}
\description{
\code{signaturePCA} performs principal component analysis on
the given data matrix and gene-signature and returns the results as an object
of class \code{prcomp}.
}
\examples{
require(Biobase)
require(breastCancerVDX)
data(vdx)
data(MSigDB)

#get the first 5000 probes of the vdx array
VDX <- readSamples(data = exprs(vdx)[1:5000,],
                   genes = fData(vdx)$Gene.symbol[1:5000])

sigPCA <- signaturePCA(MSigDB[["DOANE_BREAST_CANCER_ESR1_UP"]], VDX)

#plot using sigPlot
sigPlot(sigPCA)
}

