% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{signaturePCA}
\alias{signaturePCA}
\title{PCA based on a given gene-signature}
\usage{
signaturePCA(data, signature, center = TRUE, scale = FALSE, ...)
}
\arguments{
\item{data}{Gene expression matrix where rownames correspond to unique gene
identifiers in \code{signature} format and columns correspond to samples.}

\item{signature}{character vector with the signature's gene identifiers.}

\item{center}{a logical value indicating whether the variables should be
shifted to be zero centered. See \code{\link{prcomp}} for more details.}

\item{scale}{a logical value indicating whether the variables should be
scaled to have unit variance before the analysis takes place. See
\code{prcomp} for more details.}

\item{...}{arguments passed to \code{prcomp}.}
}
\value{
A list with class \code{prcomp}.
}
\description{
\code{signaturePCA} performs principal component analysis on
the given data matrix and gene-signature and returns the results as an object
of class \code{prcomp}.
}
\examples{

require(Biobase)
require(breastCancerVDX)
data(vdx)

VDX <- parseData(data = exprs(vdx)[5000:9000,],
                 geneIds = fData(vdx)$Gene.symbol[5000:9000])

dummySig <- c("AGR2", "SCGB1D2", "SCGB2A2")

}

