% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{signaturePCA}
\alias{signaturePCA}
\title{PCA based on a given gene-signature}
\usage{
signaturePCA(data, signature, center = TRUE, scale = FALSE, ...)
}
\arguments{
\item{data}{Gene expression matrix where rownames correspond to unique gene
identifiers in \code{signature} format and columns correspond to samples.}

\item{signature}{character vector with the signature's gene identifiers.}

\item{center}{a logical value indicating whether the variables should be
shifted to be zero centered. See \code{\link{prcomp}} for more details.}

\item{scale}{a logical value indicating whether the variables should be
scaled to have unit variance before the analysis takes place. See
\code{prcomp} for more details.}

\item{...}{arguments passed to \code{prcomp}.}
}
\value{
A list with class \code{prcomp}.
}
\description{
\code{signaturePCA} performs principal component analysis on
the given data matrix and gene-signature and returns the results as an object
of class \code{prcomp}.
}
\examples{

dummyData <- do.call(rbind, lapply(seq(0.1, 1, by = 0.1),
                     rnorm, n = 100, m = 6))
rownames(dummyData) <- paste(rep("gene", nrow(dummyData)),
                             seq(1, nrow(dummyData)), sep = "")
dummySig <- c("gene1", "gene8", "gene9", "gene10")

dummyPca <- signaturePCA(dummyData, dummySig)

}

