% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{signaturePCA}
\alias{signaturePCA}
\title{PCA based on a given gene-signature}
\usage{
signaturePCA(signature, data, center = TRUE, scale = FALSE, ...)
}
\arguments{
\item{signature}{character vector with the signature's gene identifiers}

\item{data}{Gene expression matrix where rownames correspond to unique gene
identifiers (HGNC format \link{http://www.genenames.org}) and columns
correspond to samples.}

\item{center}{a logical value indicating whether the variables should be
shifted to be zero centered. See \code{\link{prcomp}} for more details.}

\item{scale}{a logical value indicating whether the variables should be
scaled to have unit variance before the analysis takes place. See
\code{prcomp} for more details.}

\item{...}{arguments passed to \code{prcomp}.}
}
\description{
\code{signaturePCA} performs principal component analysis on
the given data matrix and gene-signature and returns the results as an object
of class \code{prcomp}.
}

