% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{signaturePCA}
\alias{signaturePCA}
\title{PCA based on a given gene-signature}
\usage{
signaturePCA(genes, data, db, center = TRUE, scale = FALSE, ...)
}
\arguments{
\item{genes}{Character vector with the gene identifiers}

\item{data}{Data matrix with gene expression values}

\item{db}{Microarray Platform}

\item{center}{a logical value indicating whether the variables should be shifted to be zero centered. See
\code{\link{prcomp}} for more details.}

\item{scale}{a logical value indicating whether the variables should be scaled to have unit variance before
the analysis takes place. See \code{prcomp} for more details.}

\item{...}{arguments passed to \code{prcomp}.}
}
\description{
\code{signaturePCA} performs principal component analysis on the given data matrix and gene-signature and
returns the results as an object of class \code{prcomp}.
}
\author{
Nikos Sidiropoulos
}

