% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{sigPlot}
\alias{sigPlot}
\title{Signature PCA plot}
\usage{
sigPlot(sigPCA, groups = NULL, text = FALSE, pcs = c(1, 2), main = "",
  palette = "Paired", ...)
}
\arguments{
\item{sigPCA}{a \code{\link{prcomp}} object, preferrably produced using
\code{\link{signaturePCA}}}

\item{groups}{optional factor variable indicating the groups that the
observations belong to.}

\item{text}{when TRUE it plots textual annotations instead of points
based on \code{groups} parameter.}

\item{pcs}{which PCs to plot. default: 1,2}

\item{main}{plot title}

\item{palette}{colorbrewer palette scheme to be used.}

\item{...}{methods passed to \code{\link{ggplot}}}
}
\value{
A ggplot2 plot
}
\description{
Plot of the selected signature's PCA using
\code{\link{ggplot2}}.
}
\examples{
sigPCA <- signaturePCA(MSigDB[["HOMOPHILIC_CELL_ADHESION"]], AML)

sigPlot(sigPCA, groups = AML_meta$karyotype)
sigPlot(sigPCA, groups = AML_meta$karyotype, text = TRUE,
        main = "HOMOPHILIC_CELL_ADHESION")
}

