% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{sigPlot}
\alias{sigPlot}
\title{Signature PCA plot}
\usage{
sigPlot(data, signature, groups = NULL, text = FALSE, pcs = c(1, 2),
  main = "", palette = "Paired", ...)
}
\arguments{
\item{data}{gene expression matrix where rownames correspond to unique gene
identifiers in \code{signature} format, and columns correspond to samples.}

\item{signature}{character vector containing the signature's gene
identifiers.}

\item{groups}{optional factor variable indicating the groups that the
observations belong to.}

\item{text}{when TRUE it plots textual annotations instead of points
based on \code{groups} parameter.}

\item{pcs}{which PCs to plot. default: 1,2}

\item{main}{plot title}

\item{palette}{colorbrewer palette scheme to be used.}

\item{...}{methods passed to \code{\link{ggplot}}}
}
\value{
A ggplot2 plot
}
\description{
Plot of the selected signature's PCA using
\code{ggplot2}.
}
\examples{

dummyData <- do.call(rbind, lapply(seq(0.1, 0.9, by = 0.1),
                     rnorm, n = 100, m = 6))

#add a row with bimodal gene expression
dummyData <- rbind(dummyData, c(rnorm(70, 6, 0.1), rnorm(30, 9, 0.1)))

rownames(dummyData) <- paste(rep("gene", nrow(dummyData)),
                             seq(1, nrow(dummyData)), sep = "")

dummySig <- c("gene1", "gene8", "gene9", "gene10")

sigPlot(dummyData, dummySig)

groups <- factor(c(rep("GroupA", 70), rep("GroupB", 30)))
sigPlot(dummyData, dummySig, groups, main = "dummySig")

}

