% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{sigPlot}
\alias{sigPlot}
\title{Signature PCA plot}
\usage{
sigPlot(sigPCA, pcs = c(1, 2), groups = NULL, text = FALSE,
  main = "Signature", palette = "Paired", ...)
}
\arguments{
\item{sigPCA}{a \code{\link{prcomp}} object, preferrably produced using
\code{\link{signaturePCA}}}

\item{pcs}{which PCs to plot. default: 1,2}

\item{groups}{optional factor variable indicating the groups that the
observations belong to.}

\item{text}{when TRUE it plots textual annotations instead of points
based on \code{groups} parameter.}

\item{main}{plot title}

\item{palette}{colorbrewer palette scheme to be used.}

\item{...}{methods passed to \code{\link{ggplot}}}
}
\description{
To do..
}
\author{
Nikos Sidiropoulos
}

